requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
eventlet==0.33.3
Werkzeug==3.0.1